requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
torch>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Deep Learning & RL
stable-baselines3>=2.0.0
gym>=0.26.0
gymnasium>=0.29.0
tensorboard>=2.13.0

# Optimization
pulp>=2.7.0
ortools>=9.6.0
cvxpy>=1.3.0

# NLP & Arabic Support
transformers>=4.30.0
arabert>=1.0.0
camel-tools>=1.5.0

# API & Web
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.0
redis>=4.6.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
hijri-converter>=2.3.0  # For Islamic calendar
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Data Validation
marshmallow>=3.19.0
cerberus>=1.3.5

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0

# Monitoring & Performance
prometheus-client>=0.17.0
sentry-sdk>=1.26.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.1.0
